requirements.txt
flask==2.3.2
groq
web3==6.11.0
python-dotenv==1.0.0